# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企业项目知识库平台",
    "description": "Phase 1 + Phase 2 完整 API:\n\n**Phase 1:**\n- 多模态文档转换管道 (PDF/Word/Excel/PPT/图片/音视频/网页 → Markdown)\n- 基础 RAG 检索 (BGE-M3 三合一 + Qdrant + RRF + Reranker)\n- MCP 工具化查询 (list/outline/search/read 四工具)\n- 引用溯源 (内联引用标记 + Markdown 章节跳转 + 源文件定位)\n\n**Phase 2:**\n- Agentic RAG (LangGraph Plan-and-Execute, 最多4步, 推理透明化)\n- Skills 体系 (文档分析/跨文档对比/项目健康/报告生成)\n- SSE 流式输出 (GET /api/v1/search/stream, 感知延迟↓80%)\n- Contextual Retrieval (chunk 上下文前缀, 检索失败↓67%)",
    "version": "3.0.0"
  },
  "paths": {
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload Document",
        "description": "上传文档 — 非阻塞，立即返回\n\n流程:\n1. 验证文件格式和大小\n2. 计算 MD5 checksum\n3. 上传源文件到 MinIO（唯一的同步 I/O，通常 <1s）\n4. 创建 DB 记录 (status=pending)\n5. 派发 Celery 异步任务（转换+索引）\n6. 立即返回 doc_id，前端轮询 GET /documents/{doc_id} 查看状态\n\n需登录；`upload_by` 可省略（默认当前用户），若填写须与 Token 用户一致。\n\n如果 MinIO 或 DB 操作失败，执行回滚：\n- MinIO 失败 → 直接返回错误\n- DB 失败 → 清理已上传的 MinIO 文件",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/documents/list/{project_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List Documents",
        "description": "列出项目中的所有文档（需为项目成员）",
        "operationId": "list_documents_api_v1_documents_list__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentInfo"
                  },
                  "title": "Response List Documents Api V1 Documents List  Project Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{doc_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document",
        "description": "获取单个文档详情（含转换状态）\n前端轮询此接口查看异步转换进度: pending → processing → completed/failed",
        "operationId": "get_document_api_v1_documents__doc_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete Document",
        "description": "删除文档及其所有索引（需编辑者或管理员）",
        "operationId": "delete_document_api_v1_documents__doc_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{doc_id}/markdown": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document Markdown",
        "description": "获取文档的 Markdown 内容",
        "operationId": "get_document_markdown_api_v1_documents__doc_id__markdown_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "upload_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload By"
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "default": "执行阶段"
          }
        },
        "type": "object",
        "required": [
          "file",
          "project_id"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
...
      "ConversionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "title": "ConversionStatus"
      },
      "DocumentInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "source_format": {
            "type": "string",
            "title": "Source Format"
          },
          "conversion_status": {
            "$ref": "#/components/schemas/ConversionStatus"
          },
          "conversion_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversion Error"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "upload_by": {
            "type": "string",
            "format": "uuid",
            "title": "Upload By"
          },
          "upload_at": {
            "type": "string",
            "format": "date-time",
            "title": "Upload At"
          },
          "converted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Converted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "original_filename",
          "source_format",
          "conversion_status",
          "upload_by",
          "upload_at"
        ],
        "title": "DocumentInfo"
      },
...
      "DocumentUploadResponse": {
        "properties": {
          "doc_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doc Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source_format": {
            "type": "string",
            "title": "Source Format"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum"
          },
          "status": {
            "$ref": "#/components/schemas/ConversionStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "project_id",
          "filename",
          "source_format",
          "checksum",
          "status",
          "message"
        ],
        "title": "DocumentUploadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}